+++ test/bigtest/helpers/index.js
@@ -2,5 +2,4 @@ export * from './setupApplication';
 export * from './prefixKeys';
 export * from './getTotalSelectedMessage';
 export * from './OverlayContainer';
-export * from './translationsProperties';
 export * from './getColumnIndexMapping';